refactor: 디렉토리 구조 변경 - UI 로직과 도메인 로직을 담당하는 모듈 분리
